other_GANS/requirements.txt
keras
git+https://www.github.com/keras-team/keras-contrib.git
matplotlib
numpy
scipy
pillow
tensorflow
#urllib
#skimage
scikit-image
#gzip
#pickle
